Update global_shell_aliases
Added 4 new flags to §7z§ (and a separate command called §7zfullsymXX§ where XX is the dictionary size) and a prefixed note using §echo§ when the command does and doesn't include symlinks:
1) -bb3 forces the most verbose output log level. While not used by the command currently, the User might add.
2) -bt forces showing execution time stats.
3) -ssw forces compressing shared files, meaning files already in use. This might be more useful on Window$ than on Linux, though, but better safe than sorry.
4) -snl in the new command forces storing symlinks as is instead of the target data.

Changed the §if ... aptitude§ one to a §for§ which loops through §aptitude§, §apt§, and §apt-get§ in that order to make an alias which is specific to that app, and once done that it exits the loop. This way §aptitude§'s not strictly needed to make use of the alias, without having to change the script yourself.

Added ShellCheck §disable§ directives to block complaints about the brace expansions expanding during definition of the alias cuz that's exactly the point.

Dunno what I was thinking with the §print_dpkg_lock_holders§ alias having that escaped single quote where it was, so I moved it.